Merged PR 340060: [Quick Visualize] fix bug: re-render while token expired
**Description:**

When using access token as string and not via our authentication class, we cannot refresh the token and "token expired" error should be thrown.
When rendering the report/quick_visualize report again, we see that nothing happens instead of throwing the same error.

**Root cause:** `_refresh_token` property is reset to default value only when using the authentication class, but when using access token as string it stays with new value `_refresh_token=True`. So when we try to re-render the widget the change detector doesn't work (there is no change in `_refresh_token` property).

So the resolution is to reset `_refresh_token` property at the beginning of `_update_access_token` function, so it will be reset in both scenarios.