Tenant choice (interactive login authentication)
It is not possible to chose a tenant during authentication process, which is an issue when you want to access another one with the same login : the default one remains selected. This feature is adding the opportunity to define a "tenant_id" argument while loging in with the interactive login authentication.