Merged PR 405811: [Jupyter Wrapper]: Implement get_saved_report functionality
Implemented the get_saved_report functionality.

Test case explanation:
We have added embed_url in create_test_report and patched it with qv object to simulate the behavior of the get_saved_report method without calling it.
The test case ensures that the get_saved_report method of the QuickVisualize class is returning a Report object with the correct Report ID.

![image.png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/b4a12249-524a-42ee-8bb5-dfb6d09c5884/pullRequests/405811/attachments/image.png)

![image (2).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/b4a12249-524a-42ee-8bb5-dfb6d09c5884/pullRequests/405811/attachments/image%20%282%29.png)

Updated the demo notebook with get_saved_report()
![image (3).png](https://dev.azure.com/powerbi/4c7b5adb-c2d0-4f18-b23c-edc4ac30f4e1/_apis/git/repositories/b4a12249-524a-42ee-8bb5-dfb6d09c5884/pullRequests/405811/attachments/image%20%283%29.png)

Related work items: #1088666, #1114001, #1121300