Merged PR 415129: [Powerbi Jupyter]: Add global auth variable
Improvement #2 of Task - [Task 1114208: [powerbi-jupyter] Improvements to support Fabric Notebook](https://dev.azure.com/powerbi/Embedded/_workitems/edit/1114208)

The Fabric team shall use the global `AUTH` to auto populate the information for a logged in user. This way the signed in user wont have to use either of `DeviceCodeLoginAuthentication` and `AuthenticationResult` to generate the auth.